requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9
jinja2==3.1.4

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Pydantic for validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Docker
docker==7.1.0
python-on-whales==0.73.0

# HTTP client for API calls
httpx==0.27.2

# YAML parsing
pyyaml==6.0.2

# Python environment
python-dotenv==1.0.1

# Logging
rich==13.9.2

# Supabase client (optional, if using their client library)
supabase==2.9.0

# Date/time utilities
python-dateutil==2.9.0
